[1] 'Redirection' and 'Display admin/user menu' after log in feature implemented and working fine but need code factoring --- [2] 'Display edit/delete(listPosts / Posts) and approve/delete (comments) buttons' feature added and working fine but need code factoring --- [3] 'Display post comment form / signal comment buttons' feature implemented  working fine but need code factoring --- [4] Css 'overflow-wrap' rules added for posts, lastposts and comments

+++ view/frontend/listPostsView.php
@@ -33,8 +33,17 @@ $id = (int) $datas['id'];
                 <button class="regularBtns"><a href="index.php?action=post&id=<?=$id?>#commentsAnchor">Commentaires</a></button>
 
         </p>
+        
+    <?php
+    // FIXME: factoriser le code avec l'affichage ou non (1)des boutons modifier / supprimer sur listPostsView et PostView (2) des boutons approuver / supprimer des com sur PostView (3) l'affichage du menu admin de template.php
+    if( (isset($_COOKIE['login']) AND $_COOKIE['login'] == 'Admin') OR  (isset($_SESSION['username']) AND $_SESSION['username'] == 'Admin'))
+    { 
+    ?>       
         <button class="adminBtns"><a href="index.php?action=manageView&id=<?=$id?>">Modifier</a></button>
         <button class="adminBtns"><a href="index.php?action=deletePost&amp;id=<?= $id?>" onclick="return confirm('Etes vous sûr?')">Supprimer</a></button>
+    <?php
+    }  
+    ?>     
     </div>
 <?php
 }
